pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "mlops_project"
version = "0.0.1"
description = "Our MLOps project that seeks to impress our teacher"
authors = [
  { name = "Emil Bruhn, Ditte Boye, Kamilla Holm, Philippa Bagger, Bror Bruland", email = "your@email.com" }
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "matplotlib>=3.8,<3.10",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch==2.6.0",
    "typer==0.15.1",
    "uvicorn>=0.36.0,<0.41",
    "wandb>=0.23.1",
    "numpy>=1.26,<2",
    "functions-framework>=3.10.0",
    "google-cloud-storage>=3.8.0",
    "hydra-core>=1.3.2",
    "torch-tb-profiler>=0.4.3",
    "tensorboard>=2.18,<2.19",
    "snakeviz>=2.2.2",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
    "pytest-cov>=6.3.0",
    "memory-profiler>=0.61.0",
    "snakeviz>=2.2.2",
    "httpx>=0.27.0",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]
